+++ transactions/constants.py
@@ -3,7 +3,7 @@ WITHDRAWAL = 2
 INTEREST = 3
 
 TRANSACTION_TYPE_CHOICES = (
-    (DEPOSIT, 'Deposit'),
-    (WITHDRAWAL, 'Withdrawal'),
-    (INTEREST, 'Interest'),
+    (DEPOSIT, "Deposit"),
+    (WITHDRAWAL, "Withdrawal"),
+    (INTEREST, "Interest"),
 )