+++ transactions/tasks.py
@@ -12,8 +12,8 @@ def calculate_interest():
     accounts = UserBankAccount.objects.filter(
         balance__gt=0,
         interest_start_date__gte=timezone.now(),
-        initial_deposit_date__isnull=False
-    ).select_related('account_type')
+        initial_deposit_date__isnull=False,
+    ).select_related("account_type")
 
     this_month = timezone.now().month
 
@@ -22,16 +22,12 @@ def calculate_interest():
 
     for account in accounts:
         if this_month in account.get_interest_calculation_months():
-            interest = account.account_type.calculate_interest(
-                account.balance
-            )
+            interest = account.account_type.calculate_interest(account.balance)
             account.balance += interest
             account.save()
 
             transaction_obj = Transaction(
-                account=account,
-                transaction_type=INTEREST,
-                amount=interest
+                account=account, transaction_type=INTEREST, amount=interest
             )
             created_transactions.append(transaction_obj)
             updated_accounts.append(account)
@@ -40,6 +36,4 @@ def calculate_interest():
         Transaction.objects.bulk_create(created_transactions)
 
     if updated_accounts:
-        UserBankAccount.objects.bulk_update(
-            updated_accounts, ['balance']
-        )
+        UserBankAccount.objects.bulk_update(updated_accounts, ["balance"])